// apps/extension/src/lib/core/types/subtitle.ts
export type LanguageCode =
  | 'zh' // Chinese
  | 'ja' // Japanese
  | 'ko' // Korean
  | 'en' // English
  | 'zh-TW' // Traditional Chinese
  | 'zh-CN' // Simplified Chinese
  | string; // Allow other language codes